+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project somewhat adheres to [Semantic Versioning](https://semver.org/sp
 - Fix rare crash while quitting Porymap.
 - Fix exported images on macOS using a different color space than in Porymap.
 - Fix some `INCBIN` statements not being parsed correctly.
+- Fix excessive logging if Porymap fails to monitor all map files.
 
 ## [6.2.0] - 2025-08-08
 ### Added